more anqfits conversion
git-svn-id: svn+ssh://astrometry.net/svn/trunk/src/astrometry@22529 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ util/fitsioutils.h
@@ -183,7 +183,8 @@ int fits_find_column(const qfits_table* table, const char* colname);
 
 int fits_get_atom_size(tfits_type type);
 
-int fits_find_table_column(const char* fn, const char* colname, int* start, int* size, int* extension);
+int fits_find_table_column(const char* fn, const char* colname,
+						   off_t* start, off_t* size, int* extension);
 
 qfits_table* fits_get_table_column(const char* fn, const char* colname, int* pcol);
